update with raw timing, cleaned up timing code
This version cleans up the timing code to make it more readable.

Adds a raw TIME and ADJUSTED time column to the output.

Fixed issues with inter-test timing. Oleg's version was trying to store the results of one test to inform future tests (aka: use integer assignment to deduct the results of integer addition later), but this wasn't actually working.

Future suggestions:
* Add integer% tests to compare manipulating int% types
* Add tests with $hex and %binary literals