VERA 0.9/R37 compatibility
Changes were required to bring it to R37 compatibility and VERA 0.9 design. Changed all relevent. Beneath the obvious changes in addressing, also the pallette needed fixing as the way of storage of the pallette and read of registers changed. Now colors are hard defined in the code. Also added some comments for better understanding of what this code does in the commander X16.