Updated two of the basic samples to VERA 0.9 definition (#105)
* Update for VERA 0.9 new registers

Changed the VPOKE's to reflect the new registers as per VERA 0.9

* Updated for VERA 0.9

Changed the addresses of Pallette vmem to be in compliance with VERA 0.9

* some typos

Forgot to add some REM commands...

* VERA 0.9/R37 compatibility

Changes were required to bring it to R37 compatibility and VERA 0.9 design. Changed all relevent. Beneath the obvious changes in addressing, also the pallette needed fixing as the way of storage of the pallette and read of registers changed. Now colors are hard defined in the code. Also added some comments for better understanding of what this code does in the commander X16.

* VERA 0.9 and R37+ compatibility

Changed all commands 1:1 with the new versions. Also added some more REMs for better understanding. Changed the address for FACLO ZP from 218 ($DA) to 199 ($C7)

* Add files via upload

A faster fractal 256 utilizing VERA 0.9 capabilities, autoincrement and writing two registers at a time to speed up fractal generation.